refactor(auth): remove unused layout components and simplify authentication routes
